Allocate H.274 film grain metadata dynamically. Saves ~779kb.
Film grain support adds a huge amount of overhead to the H264Context
structure for a feature that is rarely used. On low end devices or
pages that have lots of media this bloats memory usage rapidly.

This changes the static film grain metadata allocations to be dynamic
which reduces the H264Context size from 851808 bytes to 53444 bytes.

This is slightly different than the version upstreamed in that it
doesn't apply the hevc changes, but we don't need those in Chrome.

R=tguilbert

Bug: https://crbug.com/359358875
Change-Id: I0269324edc2d58ef5a645c4585d903c1d01a1a00
Signed-off-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/5786487
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>